pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netbox-plugin-dns-bridge"
dynamic = ["version"]
description = "A bridge between netbox-plugin-dns and your DNS infrastructure with DDNS and IXFR support."
requires-python = ">=3.10"
dependencies = ["netbox-plugin-dns>=1.5.0", "dnspython>=2.8.0"]
license = { text = "MIT" }
authors = [{ name = "Sven Luethi", email = "dev@sven.luethi.co" }]

[project.optional-dependencies]
dev = ["django-stubs", "pyright"]

[project.urls]
Homepage = "https://github.com/Suraxius/netbox-plugin-dns-bridge"
Issues = "https://github.com/Suraxius/netbox-plugin-dns-bridge/issues"

[tool.setuptools]
include-package-data = true
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
version = { attr = "netbox_dns_bridge.__version__" }
